+++ src/js/src/index.ts
@@ -1,34 +1,5 @@
-import { mount, BaseReactPyClient, ReactPyClient } from "@reactpy/client";
-import { createReconnectingWebSocket } from "./utils";
-import { ReactPyDjangoClientProps, ReactPyUrls } from "./types";
-
-export class ReactPyDjangoClient
-	extends BaseReactPyClient
-	implements ReactPyClient
-{
-	urls: ReactPyUrls;
-	socket: { current?: WebSocket };
-
-	constructor(props: ReactPyDjangoClientProps) {
-		super();
-		this.urls = props.urls;
-		this.socket = createReconnectingWebSocket({
-			readyPromise: this.ready,
-			url: this.urls.componentUrl,
-			onMessage: async ({ data }) =>
-				this.handleIncoming(JSON.parse(data)),
-			...props.reconnectOptions,
-		});
-	}
-
-	sendMessage(message: any): void {
-		this.socket.current?.send(JSON.stringify(message));
-	}
-
-	loadModule(moduleName: string) {
-		return import(`${this.urls.jsModules}/${moduleName}`);
-	}
-}
+import { mount } from "@reactpy/client";
+import { ReactPyDjangoClient } from "./client";
 
 export function mountComponent(
 	mountElement: HTMLElement,
@@ -36,6 +7,7 @@ export function mountComponent(
 	urlPrefix: string,
 	componentPath: string,
 	resolvedJsModulesPath: string,
+	reconnectStartInterval: number,
 	reconnectMaxInterval: number,
 	reconnectJitterMultiplier: number,
 	reconnectBackoffMultiplier: number,
@@ -68,7 +40,7 @@ export function mountComponent(
 		}
 	}
 
-	// Set up the client
+	// Configure a new ReactPy client
 	const client = new ReactPyDjangoClient({
 		urls: {
 			componentUrl: `${wsOrigin}/${urlPrefix}/${componentPath}`,
@@ -76,9 +48,10 @@ export function mountComponent(
 			jsModules: `${httpOrigin}/${jsModulesPath}`,
 		},
 		reconnectOptions: {
+			startInterval: reconnectStartInterval,
 			maxInterval: reconnectMaxInterval,
-			intervalJitter: reconnectJitterMultiplier,
-			backoffRate: reconnectBackoffMultiplier,
+			jitterMultiplier: reconnectJitterMultiplier,
+			backoffMultiplier: reconnectBackoffMultiplier,
 			maxRetries: reconnectMaxRetries,
 		},
 	});
