+++ src/reactpy_django/templates/reactpy/component.html
@@ -14,6 +14,7 @@
     "{{ reactpy_url_prefix }}",
     "{{ reactpy_component_path }}",
     "{{ reactpy_resolved_web_modules_path }}",
+    Number("{{ reactpy_reconnect_interval }}"),
     Number("{{ reactpy_reconnect_max_interval }}"),
     Number("{{ reactpy_reconnect_jitter_multiplier }}"),
     Number("{{ reactpy_reconnect_backoff_multiplier }}"),
